+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - SimpleDB: Use default server
 - CSS: Improve contrast of query time and dark mode
 - Editor: Display all images recognized by PHP, set their size, load them lazily
+- Designs: Show version of Adminer next to a used design not matching this version
 - Plugins: Method verifyVersion()
 - Plugins: Methods commentValue() and commentInput() (bug #1233)
 - Plugins: Show driver plugins in the list of loaded plugins